refactor(scripts): hoist ScriptUploadModal to module scope (#41)
ScriptUploadModal was defined inside the LoveTimeApp body, so its function
identity changed on every App re-render. React treated each render as a new
component type and unmounted + remounted the modal, discarding its local form
state (title, scenario, content, selected image) and re-running its thumbnail
URL.createObjectURL effect. Any App re-render while the modal was open (a poll,
a notification, a context update) silently wiped the user's in-progress form.

Move it to src/components/ScriptUploadModal.tsx with a stable module-scope
identity, threading editingScript, onClose, addCustomScript, and
updateCustomScript through props. Form state now survives App re-renders.
